Move ExpirationType and add TokenType
