+++ src/main/java/com/bukrain/share/token/TokenType.java
@@ -1,0 +1,6 @@
+package com.bukrain.share.token;
+
+public enum TokenType {
+    SINGLE_USE,
+    TIME_BASED
+}
